add users to db and refactor migrations, seeds, models

+++ db/migrations/20210429231727_addresses.ts
@@ -2,20 +2,30 @@ import { Knex } from 'knex';
 
 const TABLE_NAME = 'addresses';
 
-exports.up = (knex: Knex) => {
-  return knex.schema.createTableIfNotExists(TABLE_NAME, (table) => {
-    table.increments();
-    table.integer('contact_id');
-    table.string('address1');
-    table.string('address2');
-    table.string('address3');
-    table.string('city');
-    table.string('county');
-    table.string('postCode');
-    table.timestamps();
-  });
-};
+export async function up(knex: Knex): Promise<void> {
+  const exists = await knex.schema.hasTable(TABLE_NAME);
 
-exports.down = (knex: Knex) => {
-  return knex.schema.dropTable(TABLE_NAME);
-};
+  if (!exists) {
+    return knex.schema.createTable(TABLE_NAME, (table) => {
+      table.increments();
+      table.integer('contactId');
+      table.string('address1');
+      table.string('address2');
+      table.string('address3');
+      table.string('city');
+      table.string('county');
+      table.string('postCode');
+      table.timestamp('createdAt');
+      table.timestamp('updatedAt');
+      });
+  }
+}
+
+export async function down(knex: Knex): Promise<void> {
+  const exists = await knex.schema.hasTable(TABLE_NAME);
+
+  if (exists) {
+    return knex.schema.dropTable(TABLE_NAME);
+  }
+}
+
