+++ db/migrations/20210429231737_comments.ts
@@ -2,15 +2,25 @@ import { Knex } from 'knex';
 
 const TABLE_NAME = 'comments';
 
-exports.up = (knex: Knex) => {
-  return knex.schema.createTableIfNotExists(TABLE_NAME, (table) => {
-    table.increments();
-    table.integer('contact_id');
-    table.string('comment');
-    table.timestamps();
-  });
-};
+export async function up(knex: Knex): Promise<void> {
+  const exists = await knex.schema.hasTable(TABLE_NAME);
 
-exports.down = (knex: Knex) => {
-  return knex.schema.dropTable(TABLE_NAME);
-};
+  if (!exists) {
+    return knex.schema.createTable(TABLE_NAME, (table) => {
+      table.increments();
+      table.integer('contactId');
+      table.string('comment');
+      table.timestamp('createdAt');
+      table.timestamp('updatedAt');
+    });
+  }
+}
+
+export async function down(knex: Knex): Promise<void> {
+  const exists = await knex.schema.hasTable(TABLE_NAME);
+
+  if (exists) {
+    return knex.schema.dropTable(TABLE_NAME);
+  }
+}
+
